Create Data module, for local persistence.
I decided to create a new module for the data, avoiding dependency between features.
But this could be on the Favorites module, making it a dependency of the Detail module.
I used the Room database, as it has a good integration and support with Android, and helps with SQLite a lot.